requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
Pillow>=10.1.0
opencv-python>=4.8.0
numpy>=1.25.0
reportlab>=4.0.4
python-magic>=0.4.27
exifread>=3.0.0
requests>=2.31.0
aiohttp>=3.8.4
python-multipart>=0.0.6
